Backend/requirements.txt
Django==5.2.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
django-cors-headers
channels
daphne
channels-redis 
django-channels-jwt-auth-middleware
